requirements.txt
fastapi
pydantic
python-decouple
PyJWT
pytest
uvicorn
SQLAlchemy